x/specy/types/genesis_test.go
package types_test

import (
	"testing"

	"github.com/stretchr/testify/require"
	"specy/x/specy/types"
)

func TestGenesisState_Validate(t *testing.T) {
	for _, tc := range []struct {
		desc     string
		genState *types.GenesisState
		valid    bool
	}{
		{
			desc:     "default is valid",
			genState: types.DefaultGenesis(),
			valid:    true,
		},
		{
			desc: "valid genesis state",
			genState: &types.GenesisState{

				ExecutorList: []types.Executor{
					{
						Address: "0",
					},
					{
						Address: "1",
					},
				},
				TaskList: []types.Task{
					{
						TaskHash: "0",
					},
					{
						TaskHash: "1",
					},
				},
				// this line is used by starport scaffolding # types/genesis/validField
			},
			valid: true,
		},
		{
			desc: "duplicated executor",
			genState: &types.GenesisState{
				ExecutorList: []types.Executor{
					{
						Address: "0",
					},
					{
						Address: "0",
					},
				},
			},
			valid: false,
		},
		{
			desc: "duplicated task",
			genState: &types.GenesisState{
				TaskList: []types.Task{
					{
						TaskHash: "0",
					},
					{
						TaskHash: "0",
					},
				},
			},
			valid: false,
		},
		// this line is used by starport scaffolding # types/genesis/testcase
	} {
		t.Run(tc.desc, func(t *testing.T) {
			err := tc.genState.Validate()
			if tc.valid {
				require.NoError(t, err)
			} else {
				require.Error(t, err)
			}
		})
	}
}
